Added snapshots.
Added domain entity snapshots, useful to save replaying a long history of events. Changed get_entity_events() to be a generator. Changed stored event repository for Cassandra to page through stored events for a domain entity, so that a long history of events can be replayed (was previously limited by what could be obtained in one call, subject to timeout). Changed Cassandra schema to have entity IDs in reversed order. Changed SQL schema to have timestamp (in milliseconds) so that time constraints can be queried.