Added event sourced generalized suffix trees.
With a generalized suffix tree a number of strings can be added, each with an unique string ID, by using the add_string() method. When the find_strings() method, which takes a substring argument, is used to query the suffix tree it returns a list of string IDs for strings that match the given substring. In this implementation, the generalized suffix tree, the nodes and the edges, are implemented as persistent event sourced entities which means the suffix tree effectively functions as a search index.