Fixed suffixed tree so strings can be added after being retrieved from a repo.
Now it has a node and an edge repo, to use with when it a suffix tree is retrieved from its repo it doesn't immediately have any existing nodes and edges in its internal cache, and needs to get them. Possible the methods requiring the repos shoud become domain services rather than entity operations.