Added snapshot strategy.
Added snapshot strategy as a argument in constructor of event sourced repos, so that the application can control how snapshots of event sourced entities are created and retrieved. Added class EventSourcedSnapshotStrategy to encapsulate the already existing way of creating and retrieving event sourced snapshots (ie snapshots as domain events, persisted in the stored events table). Removed the previously added code that allowed event players to be customised (design reject).