Made expected version checking optional (was compulsory).
Added param to application class constructor to allow expected version checking to be switched on (so that it isn't by default, and there doesn't break existing applications which will not have any expected versions). Versions are written still, so the new entity version table will need to be created in your migration script.