Added methods to get all domain events in the application.
Added method to event store to get all domain events in the
application, and a method to active record strategy to get
all the items in all sequences. This allows (in SQLAlchemy)
for the entire application event stream to be replayed in
the correct order. In Cassandra, the events cannot be returned
in order, but nevertheless this will allow all the events to
be recovered (if necessary, depending on the number of events
and the capacity of a sequence, the events could be projected
into a single sequence in a different event store).