Split out user guide into several files.
Reworked examples in user guide to use library classes
more, previously reused code from previous sections
but that was broken when the file was split. Also
refactored to harmonise library examples with user
guide. Copy still needs smoothing out after the upheaval,
and the examples might need to be adjusted.