Improved process doc.
Also changed system to run with a shared database by default, so
it's more similar to multiprocessing with a shared database.
Multiple databases is a more advanced configuration, and this
will avoid some kinds of errors (e.g. getting record conflicts
with shared database because of a programming error and thinking
it has something to do with multiprocessing....).