Changed "use_direct_query_if_available" to be False by default, and to be a constructor arg of both ProcessApplication and SystemRunner.
This was using a "fast" path to get notifications, but this path
doesn't page and so effectively locks up when there are too many
notifications, which there can be when reprocessing a large
notification log. Also, since the notification log now gets the
first section it needs directly, the "fast" path isn't faster
than the normal path anymore.