Improved record manager test cases, and added missing things.
Basically, I was thinking about using the 'set_notification_ids'
feature of application classes more, so that notifications
can be put into shared memory without needing to query the
database to get the notification IDs, and felt like the
testing of the "insert values" mode needed to be covered
better, in particular covered by the record manager tests
rather than only exercised by the application test cases.
In adding base test cases for these things, I was able to
test the record managers in a standard way, and then it
made sense to implement previously missing support for
these aspects in some record managers, and also improve
the tests so that now I feel more confident in this aspect.

- Previously, the "insert values" mode of record managers
  wasn't covered by the record manager tests.
- Django and Popo record managers didn't previously support
  "insert values" mode, which allows for notification IDs to
  be set in the application, and "non-notifiable" domain
  events (where the domain events aren't present in the
  notification log).
- Django, Popo, and SQLAlchemy record managers didn't
  previously reject non-integer values for notification IDs
  (which can cause problems).
- SQLAlchemy record manager wasn't using compiled insert
  statements where it could (ie when there are no ORM objects
  to insert). Still might be possible to use them with ORM
  objects, but it seems to need the connection rather than
  transaction and ORM objects seem to need the transaction
  rather than connection interface?
- Previously, test cases were using methods on subclasses
  of record manager but the factory method for the record
  manager was type-hinted with the superclass, so couldn't
  navigate to the method-under-test.
- Previously, the tracking record aspect of record managers
  wasn't being tested in the record manager test cases.