Adjusted order of mutate() method, so version is not assigned before apply() method is called, in case apply() raises an error.
Fixed event decorator to use setter arg name in event args
when event decorator is used above setter decorator (was using
setter method name, which caused an exception when arg name
is different from name of property which it was in the tests).

Fixed some mypy issues (e.g. use of event decorator on __init__).