Changed postgres.py to use connection pool.
This became necessary for the case of firing
up a new thread to do something once and then
repeating. But it also helps minimise the number
of connections used that need to be created.

This required reworking transactions to be
issued by connections, which makes lots of
sense; because we needed to make sure we
have the same connection when preparing
statements, and preparing statements needed
to be done in a separate transaction. In
other words, the same connection needed to
be used for more than one transaction and
introducing a connection pool rather than
getting the connection for a thread messed
up the way it used to work.

Still needs more work: to use a queue to wait
for connections to be return when pool is
exhausted rather than raise "pool exhausted"
error. Also cycling over connections in pool
in case they are all closed needs more attention.