Added per-aggregate locking around fast-forwarding of cached aggregates.
Because fast-forwarding isn't thread-safe. Slightly complicated because we probably don't want to accumulate vast numbers of locks (hence using the LRUCache for totally disused locks), but we also don't want to risk having two locks for the same aggregate (hence counting number of users of a lock that is in use). And counting users and moving between cached to in-use collections isn't thread-safe (hence the "locks lock").