Improved test for "fairness" in getting connection from pool.
Was a bit flakey, because running threads introduces some randomness, so threads were nipping in front of others to make their request, which was indistinguishable from unfair allocation of connections. However, by slowing things down, requests to get a connection can be made in a stable order, and so then we can see that connections are issued to requests fairly (in the order the connection was requested). If one thread makes a request but a later request is given a connection "unfairly" first, so requests are stalled excessively, the deadline will be exceeded and the test will fail. But it might fail anyway if the entire Python process is somehow blocked (so it's probably still flakey in that respect).